+++ onedal/ensemble/tests/test_random_forest.py
@@ -22,19 +22,9 @@ from daal4py.sklearn._utils import daal_check_version
 from onedal.ensemble import RandomForestClassifier, RandomForestRegressor
 from onedal.tests.utils._device_selection import get_queues
 
-hparam_values = [
-    (None, None, None, None),
-    (8, 100, 32, 0.3),
-    (16, 100, 32, 0.3),
-    (32, 100, 32, 0.3),
-    (64, 10, 32, 0.1),
-    (128, 100, 1000, 1.0),
-]
-
 
 @pytest.mark.parametrize("queue", get_queues())
-@pytest.mark.parametrize("block, trees, rows, scale", hparam_values)
-def test_rf_classifier(queue, block, trees, rows, scale):
+def test_rf_classifier(queue):
     X, y = make_classification(
         n_samples=100,
         n_features=4,
@@ -44,12 +34,6 @@ def test_rf_classifier(queue, block, trees, rows, scale):
         shuffle=False,
     )
     rf = RandomForestClassifier(max_depth=2, random_state=0).fit(X, y, queue=queue)
-    if daal_check_version((2024, "P", 300)) and block is not None:
-        hparams = rf.get_hyperparameters("infer")
-        hparams.block_size = block
-        hparams.min_trees_for_threading = trees
-        hparams.min_number_of_rows_for_vect_seq_compute = rows
-        hparams.scale_factor_for_vect_parallel_compute = scale
     assert_allclose([1], rf.predict([[0, 0, 0, 0]], queue=queue))
